+++ src/main/java/Application/HandlerAssembler.java
@@ -1,3 +1,5 @@
+package Application;
+
 import java.util.Arrays;
 import java.util.List;
