+++ src/main/java/Application/RedirectHandler.java
@@ -1,3 +1,9 @@
+package Application;
+
+import Server.RequestParser;
+import Server.ResponseBuilder;
+import Server.Response;
+
 import java.util.Arrays;
 import java.util.List;
