+++ src/main/java/Server/CommunicationChannel.java
@@ -1,3 +1,5 @@
+package Server;
+
 import java.io.InputStream;
 import java.io.OutputStream;
