+++ src/main/java/Server/RequestParser.java
@@ -1,3 +1,5 @@
+package Server;
+
 public class RequestParser {
 
     private String request;
