+++ src/main/java/Server/RequestReader.java
@@ -1,3 +1,5 @@
+package Server;
+
 import java.io.IOException;
 import java.io.InputStream;
