+++ src/main/java/Server/ResponseBuilder.java
@@ -1,3 +1,5 @@
+package Server;
+
 public class ResponseBuilder {
 
     private String statusCode;
