+++ src/main/java/Server/ResponseHandler.java
@@ -1,3 +1,5 @@
+package Server;
+
 import java.io.IOException;
 import java.io.OutputStream;
