+++ src/main/java/Server/ServerStatus.java
@@ -1,3 +1,5 @@
+package Server;
+
 public interface ServerStatus {
     boolean acceptConnections();
 }
