+++ src/main/java/Server/SocketManager.java
@@ -1,3 +1,5 @@
+package Server;
+
 import java.io.IOException;
 
 public interface SocketManager {
