+++ src/main/java/Server/WebServer.java
@@ -1,3 +1,8 @@
+package Server;
+
+import Application.Handler;
+import Application.HandlerAssembler;
+
 import java.io.*;
 import java.util.List;
