+++ src/test/java/GetWithBodyHandlerTest.java
@@ -1,3 +1,5 @@
+import Application.GetWithBodyHandler;
+import Server.RequestParser;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
