+++ src/test/java/MockServerCommunicationChannel.java
@@ -1,3 +1,5 @@
+import Server.CommunicationChannel;
+
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
