+++ src/test/java/NotFoundResourceHandlerTest.java
@@ -1,3 +1,5 @@
+import Application.NotFoundResourceHandler;
+import Server.RequestParser;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
