+++ src/test/java/RequestParserTest.java
@@ -1,3 +1,4 @@
+import Server.RequestParser;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
