+++ src/test/java/RequestReaderTest.java
@@ -1,3 +1,4 @@
+import Server.RequestReader;
 import org.junit.Assert;
 import org.junit.Test;
