+++ src/test/java/SimpleGetHandlerTest.java
@@ -1,3 +1,5 @@
+import Application.SimpleGetHandler;
+import Server.RequestParser;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
