+++ src/test/java/WebServerStatusTest.java
@@ -1,3 +1,4 @@
+import Server.WebServerStatus;
 import org.junit.Assert;
 import org.junit.Test;
