+++ src/test/java/WebServerTest.java
@@ -1,3 +1,4 @@
+import Server.*;
 import org.junit.Assert;
 import org.junit.Test;
